pi/requirements.txt
RPi.GPIO
requests
configobj